Configure bundler 3 build to process React code
Run bundler 3 through the project to a make a single main.js file to feed to the
fragments cli.

Because the fragments cli does not allow configuration of the "src" folder, I
have had to create a "react-src" folder which, when processed by bundler 3
puts everything in "src" for the fragments cli to continue.

Note that you need a development version of bundler 3 to build this (as it has
not been released yet, because it this is WIP).

For the record, the main.js file generated by bundler 3 looks like [this](https://gist.github.com/izaera/3cf1ada112f1b23a8dc88110d956501a).